fix: base update with associations
